Add sketch folder to include directories
Because everything is built in another directory than the sketch folder, it is impossible to share common code between sketches or other C++ projects when all of them are in the same repository. Even with the library for Arduino, it would be strange to use one in another C++ project. This commit aims to make the share possible by adding the possibility to specify relative include paths to the sketch file.